Instead of putting it all under __init__.py, move main.py to proxy.py
