+++ proxy/__main__.py
@@ -7,7 +7,7 @@
     :copyright: (c) 2013-present by Abhinav Singh and contributors.
     :license: BSD, see LICENSE for more details.
 """
-from . import entry_point
+from .proxy import entry_point
 
 if __name__ == '__main__':
     entry_point()
