+++ requirements.txt
@@ -1,0 +1,1 @@
+typing-extensions==3.7.4.1
